requirements.txt
Flask>=2.0
mysql-connector-python>=8.0
livekit-agents[deepgram,openai,cartesia,silero,turn-detector,google]~=1.0
livekit-plugins-noise-cancellation~=0.2
python-dotenv
